Reduce mem used in RunningAverage class
An array is created of this type for the entire DEM grid.
Before this commit, an instance took up 16 bytes. This lead
to out-of-memory errors for bench1 and bench2.

We can get the size down to 8 bytes by using an int instead of
a long. This should be fine, as there likely will not be more
than a million or two values per grid point. If that changes,
this may need to be re-evaluated.

+++ Typedefine.hpp
@@ -1132,7 +1132,7 @@ class RunningAverage {
 
     private:
         float v;
-        long n;
+        int n;
 
 };
